Mycroft_Framework/rl_portfolio/requirements.txt
# Core RL
torch>=2.0.0
stable-baselines3>=2.0.0
gymnasium>=0.29.0
sb3-contrib>=2.0.0

# Financial Data
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0
plotly>=5.14.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
scikit-learn>=1.3.0
